chore: move reinitialiseDB out of reinit
