+++ test/fields/collections/Lexical/e2e.spec.ts
@@ -5,7 +5,7 @@ import type { SerializedEditorState, SerializedParagraphNode, SerializedTextNode
 
 import { expect, test } from '@playwright/test'
 import { initPayloadE2ENoConfig } from 'helpers/initPayloadE2ENoConfig.js'
-import { reInitializeDB } from 'helpers/reInit.js'
+import { reInitializeDB } from 'helpers/reInitializeDB.js'
 import path from 'path'
 import { wait } from 'payload/utilities'
 import { fileURLToPath } from 'url'
