+++ test/helpers/reInitializeDB.ts
@@ -1,0 +1,22 @@
+export const path = '/re-initialize'
+
+export const reInitializeDB = async ({
+  serverURL,
+  snapshotKey,
+  uploadsDir,
+}: {
+  serverURL: string
+  snapshotKey: string
+  uploadsDir?: string
+}) => {
+  await fetch(`${serverURL}/api${path}`, {
+    method: 'post',
+    body: JSON.stringify({
+      snapshotKey,
+      uploadsDir,
+    }),
+    headers: {
+      'Content-Type': 'application/json',
+    },
+  })
+}
